10주차 - POV
수동 POV 방식으로 맵 구성

+++ Assets/Scripts/Node.cs
@@ -4,16 +4,16 @@ using UnityEngine;
 
 public class Node : MonoBehaviour
 { 
-    private Dictionary<Node, int> _dic;
+    private Dictionary<Node, float> _dic;
     public bool isVisited;
 
-    public Dictionary<Node, int> nodeDic
+    public Dictionary<Node, float> nodeDic
     {
         get
         {
             if (_dic == null)
             {
-                _dic = new Dictionary<Node, int>();
+                _dic = new Dictionary<Node, float>();
             }
             return _dic;
         }
